Move Puppeteer from cdp to common
