+++ packages/puppeteer-core/src/cdp/Puppeteer.ts
@@ -15,16 +15,10 @@
  */
 
 import type {Browser} from '../api/Browser.js';
-import type {ConnectionTransport} from '../common/ConnectionTransport.js';
-import {
-  type CustomQueryHandler,
-  customQueryHandlers,
-} from '../common/CustomQueryHandler.js';
+import {type CustomQueryHandler, customQueryHandlers,} from '../common/CustomQueryHandler.js';
 
-import {
-  type BrowserConnectOptions,
-  _connectToCdpBrowser,
-} from './BrowserConnector.js';
+import {_connectToCdpBrowser,} from './BrowserConnector.js';
+import {ConnectOptions} from './ConnectOptions.js';
 
 /**
  * Settings that are common to the Puppeteer class, regardless of environment.
@@ -33,20 +27,6 @@ import {
  */
 export interface CommonPuppeteerSettings {
   isPuppeteerCore: boolean;
-}
-/**
- * @public
- */
-export interface ConnectOptions extends BrowserConnectOptions {
-  browserWSEndpoint?: string;
-  browserURL?: string;
-  transport?: ConnectionTransport;
-  /**
-   * Headers to use for the web socket connection.
-   * @remarks
-   * Only works in the Node.js environment.
-   */
-  headers?: Record<string, string>;
 }
 
 /**
