+++ packages/puppeteer-core/src/node/PuppeteerNode.ts
@@ -26,7 +26,6 @@ import type {Browser} from '../api/Browser.js';
 import type {BrowserConnectOptions} from '../cdp/BrowserConnector.js';
 import {
   type CommonPuppeteerSettings,
-  type ConnectOptions,
   Puppeteer,
 } from '../cdp/Puppeteer.js';
 import type {Configuration} from '../common/Configuration.js';
@@ -41,6 +40,7 @@ import type {
   LaunchOptions,
 } from './LaunchOptions.js';
 import type {ProductLauncher} from './ProductLauncher.js';
+import {ConnectOptions} from '../cdp/ConnectOptions.js';
 
 /**
  * @public
